- Merge branch 'APPLINK-9800-RepairUnitTestQNX-KKolodiy' into develop

+++ src/components/policy/src/policy/qdb_wrapper/include/qdb_wrapper/sql_query.h
@@ -241,7 +241,7 @@ class SQLQuery {
    */
   Error error_;
 
-  int SetBinds();
+  uint8_t SetBinds();
   bool Result();
 };
